+++ apps/erlangbridge/test/vscode_connection_SUITE_data/attach_target.erl
@@ -1,0 +1,5 @@
+-module(attach_target).
+-export([add/2]).
+
+add(A, B) ->
+    A + B.